ai/requirements.txt
# AI/ML Dependencies for Pill Classification System

# Deep Learning Frameworks
ultralytics>=8.0.0  # YOLOv8
torch>=1.13.0
torchvision>=0.14.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data Processing
numpy>=1.21.0
pandas>=1.5.0

# Model Training & Evaluation
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.11.0

# API & Server Integration
Flask>=2.2.0
requests>=2.28.0

# Utilities
tqdm>=4.64.0
argparse